+++ src/main.ts
@@ -1,3 +1,16 @@
-import program from "./cli";
+import { Command } from "commander";
+import { getUserName } from "./user";
+import { startChat } from "./chat";
 
-program.parse(process.argv);
+const program = new Command();
+
+program
+  .version("1.0.0")
+  .description("CLI Chat Application with LLM")
+  .action(async () => {
+    const userName = getUserName();
+    console.log(`Hello, ${userName}! Welcome to the CLI application.`);
+    await startChat();
+  });
+
+  program.parse(process.argv);
